requirements.txt
django
djangorestframework
pylint
dj-rest-auth[with_social]
django-filter
django-treebeard
coreapi
pyyaml
pipenv
drf-spectacular
drf-nested-routers
